made userid an int, which ended up touching a lot
Former-commit-id: 335c86dc7e2cb36ca6badbdefdefa5fab7547602 [formerly 60ca7913af5e236a6d015e3bd761629d75930d73]
Former-commit-id: a458a8360b7c48e57b6d594f798c50f0951f6aa8

+++ src/mitll/langtest/shared/monitoring/Session.java
@@ -38,7 +38,7 @@ public class Session implements IsSerializable, SetScore {
    * @param userid
    * @param timestamp
    */
-  public Session(int id, long userid, long timestamp) {
+  public Session(int id, int userid, long timestamp) {
     this.userid = userid;
     this.id = id;
     this.timestamp = timestamp;
